refactor: 추상 클래스 활용하여 코드 중복 제거
